refactor: xss filter is config package -> security package (#83)
